+++ devel.R
@@ -24,5 +24,3 @@ fakemake::visualize(ml)
 
 
 packager::create("../cuwintest")
-
-